requirements.txt
Django==5.2.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.5.1
django-encrypted-model-fields==0.6.5
cryptography==46.0.2
mysqlclient==2.2.4